requirements.txt
langchain
langchain-openai
openai
streamlit
faiss-cpu
tiktoken
python-dotenv
pypdf
unstructured
langchain-community
pandas
chromadb
pdfplumber